Response to PR feedback.
